feature: 📦 ready for production deployment
This commit set up the api to run on podman production environment.
The Dockerfile build a image with the webserver (meinheld) and a reverse proxy
(nginx) -Thx tiangolo for his amazing work
(https://github.com/tiangolo/meinheld-gunicorn-flask-docker). The only aditional
required part is the postgres database.

There was a lot of unnecessary files, so those were removed.