refactor: 🔪 switch flask restful for restx
For catching all Flask-JWT-Extended errors with ease, was needed to
switch Flask-Restful for Flask-Restx. There was no easy way of doing
with the old plugin. Another reason was that Flask-Restful mainetance
is stalled.

The problem is that in Flask-Restx the resource class works diferent.
That's why the endpoints/*.py __init__() method was modified.

refactor: 🐵 refactor agenda and core module